@@ -1,18 +1,20 @@
 name: native_assets_builder
 description: >-
   This package is the backend that invokes top-level `build.dart` scripts.
-version: 0.5.0
+version: 0.6.0-wip
 repository: https://github.com/dart-lang/native/tree/main/pkgs/native_assets_builder
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
+
+publish_to: none
 
 dependencies:
   graphs: ^2.3.1
   logging: ^1.2.0
-  native_assets_cli: ^0.4.2
-  # native_assets_cli:
-  #   path: ../native_assets_cli/
+  # native_assets_cli: ^0.4.2
+  native_assets_cli:
+    path: ../native_assets_cli/
   package_config: ^2.1.0
   yaml: ^3.1.2
   yaml_edit: ^2.1.0
