@@ -5,10 +5,9 @@ version: 0.1.0
 publish_to: none
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
-  cli_config: ^0.1.1
   cyclic_package_2:
     path: ../cyclic_package_2
   native_assets_cli:
