@@ -5,14 +5,12 @@ version: 0.1.0
 publish_to: none
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
-  cli_config: ^0.1.1
   logging: ^1.1.1
   native_assets_cli:
     path: ../../../native_assets_cli/
-  # native_toolchain_c: ^0.3.4+1
   native_toolchain_c:
     path: ../../../native_toolchain_c/
 
