@@ -4,8 +4,6 @@
 
 import 'package:native_assets_cli/native_assets_cli.dart';
 
-void main(List<String> args) async {
-  final buildConfig = await BuildConfig.fromArgs(args);
-  final buildOutput = BuildOutput();
-  await buildOutput.writeToFile(outDir: buildConfig.outDir);
+void main(List<String> arguments) async {
+  await build(arguments, (config, output) async {});
 }
