@@ -2,20 +2,20 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:cli_config/cli_config.dart';
 import 'package:native_assets_cli/native_assets_cli.dart';
 
 void main(List<String> args) async {
-  final config = await Config.fromArgs(args: args);
-  final buildConfig = BuildConfig.fromConfig(config);
+  final buildConfig = BuildConfig(args);
   if (!buildConfig.dryRun) {
-    final metadata =
-        buildConfig.dependencyMetadata!['package_with_metadata']!.metadata;
-    final someValue = metadata['some_key'];
+    final someValue =
+        buildConfig.metadatum('package_with_metadata', 'some_key');
     assert(someValue != null);
-    final someInt = metadata['some_int'];
+    final someInt = buildConfig.metadatum('package_with_metadata', 'some_int');
     assert(someInt != null);
-    print(metadata);
+    print({
+      'some_int': someInt,
+      'some_key': someValue,
+    });
   } else {
     print('meta data not available in dry run');
   }
