@@ -6,13 +6,21 @@
 /// native assets CLI, e.g. a `build.dart` script.
 library native_assets_cli;
 
-export 'src/api/asset.dart';
-export 'src/api/build_config.dart';
-export 'src/api/build_mode.dart';
-export 'src/api/build_output.dart';
-export 'src/api/dependencies.dart';
-export 'src/api/ios_sdk.dart';
-export 'src/api/link_mode.dart';
-export 'src/api/link_mode_preference.dart';
-export 'src/api/metadata.dart';
-export 'src/api/target.dart';
+export 'src/api/architecture.dart' show Architecture;
+export 'src/api/asset.dart'
+    show
+        Asset,
+        DynamicLoadingBundled,
+        DynamicLoadingSystem,
+        LinkMode,
+        LookupInExecutable,
+        LookupInProcess,
+        NativeCodeAsset,
+        StaticLinking;
+export 'src/api/build.dart';
+export 'src/api/build_config.dart' show BuildConfig, CCompilerConfig;
+export 'src/api/build_mode.dart' show BuildMode;
+export 'src/api/build_output.dart' show BuildOutput;
+export 'src/api/ios_sdk.dart' show IOSSdk;
+export 'src/api/link_mode_preference.dart' show LinkModePreference;
+export 'src/api/os.dart' show OS;
