@@ -2,26 +2,40 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:cli_config/cli_config.dart';
+import 'package:collection/collection.dart';
+import 'package:crypto/crypto.dart';
 import 'package:pub_semver/pub_semver.dart';
 
-import '../model/build_config.dart' as model;
-import '../model/build_mode.dart' as model;
-import '../model/ios_sdk.dart' as model;
-import '../model/link_mode_preference.dart' as model;
-import '../model/target.dart' as model;
+import '../model/metadata.dart';
+import '../model/target.dart';
+import '../utils/map.dart';
+import '../utils/yaml.dart';
+import 'architecture.dart';
+import 'asset.dart';
+import 'build.dart';
 import 'build_mode.dart';
 import 'ios_sdk.dart';
 import 'link_mode_preference.dart';
-import 'metadata.dart';
-import 'target.dart';
+import 'os.dart';
 
-abstract class BuildConfig {
-  /// The folder in which all output and intermediate artifacts should be
+part 'c_compiler_config.dart';
+part '../model/build_config.dart';
+part '../model/c_compiler_config.dart';
+
+/// The configuration for a `build.dart` invocation.
+///
+/// A package can choose to have a toplevel `build.dart` script. If such a
+/// script exists, it will be automatically run, by the Flutter and Dart SDK
+/// tools. The script will then be run with specific commandline arguments,
+/// which [BuildConfig] can parse and provide more convenient access to.
+abstract final class BuildConfig {
+  /// The directory in which all output and intermediate artifacts should be
   /// placed.
-  Uri get outDir;
+  Uri get outputDirectory;
 
   /// The name of the package the native assets are built for.
   String get packageName;
@@ -32,126 +46,92 @@ abstract class BuildConfig {
   /// dependency of another. For this it is convenient to know the packageRoot.
   Uri get packageRoot;
 
-  /// The target being compiled for.
-  ///
-  /// Not available in [dryRun].
-  Target get target;
-
   /// The architecture being compiled for.
   ///
-  /// Not available in [dryRun].
-  Architecture get targetArchitecture;
+  /// Not specified (`null`) during a [dryRun].
+  Architecture? get targetArchitecture;
 
   /// The operating system being compiled for.
-  OS get targetOs;
+  OS get targetOS;
 
   /// When compiling for iOS, whether to target device or simulator.
   ///
-  /// Required when [targetOs] equals [OS.iOS].
+  /// Not available if [targetOS] is [OS.iOS]. Will throw a [StateError] if
+  /// accessed during a [dryRun].
   ///
-  /// Not available in [dryRun].s
-  IOSSdk? get targetIOSSdk;
+  /// Not available during a [dryRun]. Will throw a [StateError] if accessed
+  /// during a [dryRun].
+  IOSSdk get targetIOSSdk;
 
   /// When compiling for Android, the minimum Android SDK API version to that
   /// the compiled code will be compatible with.
   ///
-  /// Required when [targetOs] equals [OS.android].
+  /// Required when [targetOS] equals [OS.android].
   ///
-  /// Not available in [dryRun].
+  /// Not available during a [dryRun]. Will throw a [StateError] if accessed
+  /// during a [dryRun].
   ///
   /// For more information about the Android API version, refer to
   /// [`minSdkVersion`](https://developer.android.com/ndk/guides/sdk-versions#minsdkversion)
   /// in the Android documentation.
   int? get targetAndroidNdkApi;
 
-  /// Preferred linkMode method for library.
+  /// The preferred [LinkMode] method for [NativeCodeAsset]s.
   LinkModePreference get linkModePreference;
 
-  /// Metadata from direct dependencies.
+  /// Metadata from a direct dependency.
   ///
-  /// The key in the map is the package name of the dependency.
+  /// The [packageName] of is the package name of the direct dependency.
   ///
-  /// The key in the nested map is the key for the metadata from the dependency.
+  /// Returns `null` if metadata was not provided.
   ///
-  /// Not available in [dryRun].
-  Map<String, Metadata>? get dependencyMetadata;
+  /// Not available during a [dryRun]. Will throw a [StateError] if accessed
+  /// during a [dryRun].
+  Object? metadatum(String packageName, String key);
 
   /// The configuration for invoking the C compiler.
   ///
-  /// Not available in [dryRun].
+  /// Not available during a [dryRun]. Will throw a [StateError] if accessed
+  /// during a [dryRun].
   CCompilerConfig get cCompiler;
 
-  /// Don't run the build, only report the native assets produced.
+  /// Whether this run is a dry-run, which doesn't build anything.
+  ///
+  /// A dry-run only reports information about which assets a build would
+  /// create, but doesn't actually create files.
   bool get dryRun;
 
-  /// The build mode that the code should be compiled in.
+  /// The [BuildMode] that the code should be compiled in.
   ///
-  /// Not available in [dryRun].
+  /// Currently [BuildMode.debug] and [BuildMode.release] are the only modes.
+  ///
+  /// Not available during a [dryRun]. Will throw a [StateError] if accessed
+  /// during a [dryRun].
   BuildMode get buildMode;
 
-  /// The underlying config.
+  /// The asset types that the invoker of this build supports.
   ///
-  /// Can be used for easier access to values on [dependencyMetadata].
-  Config get config;
+  /// Currently known values:
+  /// * [NativeCodeAsset.type]
+  /// * [DataAsset.type]
+  Iterable<String> get supportedAssetTypes;
 
   /// The version of [BuildConfig].
   ///
-  /// This class is used in the protocol between the Dart and Flutter SDKs
+  /// The build config is used in the protocol between the Dart and Flutter SDKs
   /// and packages through `build.dart` invocations.
   ///
-  /// If we ever were to make breaking changes, it would be useful to give
-  /// proper error messages rather than just fail to parse the YAML
-  /// representation in the protocol.
-  static Version get version => model.BuildConfig.version;
-
-  factory BuildConfig({
-    required Uri outDir,
-    required String packageName,
-    required Uri packageRoot,
-    required BuildMode buildMode,
-    required Architecture targetArchitecture,
-    required OS targetOs,
-    IOSSdk? targetIOSSdk,
-    int? targetAndroidNdkApi,
-    CCompilerConfig? cCompiler,
-    required LinkModePreference linkModePreference,
-    Map<String, Metadata>? dependencyMetadata,
-  }) =>
-      model.BuildConfig(
-        outDir: outDir,
-        packageName: packageName,
-        packageRoot: packageRoot,
-        buildMode: buildMode as model.BuildMode,
-        targetArchitecture: targetArchitecture as model.Architecture,
-        targetOs: targetOs as model.OS,
-        targetIOSSdk: targetIOSSdk as model.IOSSdk?,
-        targetAndroidNdkApi: targetAndroidNdkApi,
-        cCompiler: cCompiler as model.CCompilerConfig?,
-        linkModePreference: linkModePreference as model.LinkModePreference,
-        dependencyMetadata: dependencyMetadata?.cast(),
-      );
-
-  factory BuildConfig.dryRun({
-    required Uri outDir,
-    required String packageName,
-    required Uri packageRoot,
-    required OS targetOs,
-    required LinkModePreference linkModePreference,
-  }) =>
-      model.BuildConfig.dryRun(
-        outDir: outDir,
-        packageName: packageName,
-        packageRoot: packageRoot,
-        targetOs: targetOs as model.OS,
-        linkModePreference: linkModePreference as model.LinkModePreference,
-      );
-
-  factory BuildConfig.fromConfig(Config config) =>
-      model.BuildConfig.fromConfig(config);
+  /// We're trying to avoid breaking changes. However, in the case that we have
+  /// to, the major version mismatch between the Dart or Flutter SDK and a
+  /// `build.dart` script will lead to a nice error message.
+  static Version get latestVersion => BuildConfigImpl.latestVersion;
 
   /// Constructs a config by parsing CLI arguments and loading the config file.
   ///
-  /// The [args] must be commandline arguments.
+  /// `build.dart` hooks will most likely use [build] instead of this
+  /// constructor.
+  ///
+  /// The [arguments] must be commandline arguments.
   ///
   /// If provided, [environment] must be a map containing environment variables.
   /// If not provided, [environment] defaults to [Platform.environment].
@@ -161,39 +141,83 @@ abstract class BuildConfig {
   /// If not provided, [workingDirectory] defaults to [Directory.current].
   ///
   /// This async constructor is intended to be used directly in CLI files.
-  static Future<BuildConfig> fromArgs(
-    List<String> args, {
+  factory BuildConfig(
+    List<String> arguments, {
     Map<String, String>? environment,
     Uri? workingDirectory,
   }) =>
-      model.BuildConfig.fromArgs(
-        args,
+      BuildConfigImpl.fromArguments(
+        arguments,
         environment: environment,
         workingDirectory: workingDirectory,
       );
-}
 
-abstract class CCompilerConfig {
-  /// Path to a C compiler.
-  Uri? get cc;
+  /// Constructs a config for a non-dry run by providing values for each field.
+  ///
+  /// `build.dart` hooks will most likely use [build] instead of this
+  /// constructor. However, for unit testing code which consumes a
+  /// [BuildConfig], this constructor facilitates easy construction.
+  ///
+  /// For the documentation of the parameters, see the equally named fields.
+  ///
+  /// Parameter [dependencyMetadata] must be a nested map `{'packageName' :
+  /// {'key' : 'value'}}` where `packageName` and `key` correspond to the
+  /// parameters in [metadatum].
+  factory BuildConfig.build({
+    required Uri outputDirectory,
+    required String packageName,
+    required Uri packageRoot,
+    required BuildMode buildMode,
+    required Architecture targetArchitecture,
+    required OS targetOS,
+    IOSSdk? targetIOSSdk,
+    int? targetAndroidNdkApi,
+    CCompilerConfig? cCompiler,
+    required LinkModePreference linkModePreference,
+    Map<String, Map<String, Object>>? dependencyMetadata,
+    Iterable<String>? supportedAssetTypes,
+  }) =>
+      BuildConfigImpl(
+        outDir: outputDirectory,
+        packageName: packageName,
+        packageRoot: packageRoot,
+        buildMode: buildMode as BuildModeImpl,
+        targetArchitecture: targetArchitecture as ArchitectureImpl,
+        targetOS: targetOS as OSImpl,
+        targetIOSSdk: targetIOSSdk as IOSSdkImpl?,
+        targetAndroidNdkApi: targetAndroidNdkApi,
+        cCompiler: cCompiler as CCompilerConfigImpl?,
+        linkModePreference: linkModePreference as LinkModePreferenceImpl,
+        dependencyMetadata: dependencyMetadata != null
+            ? {
+                for (final entry in dependencyMetadata.entries)
+                  entry.key: Metadata(entry.value.cast())
+              }
+            : {},
+        supportedAssetTypes: supportedAssetTypes,
+      );
 
-  /// Path to a native linker.
-  Uri? get ld;
-
-  /// Path to a native archiver.
-  Uri? get ar;
-
-  /// Path to script that sets environment variables for [cc], [ld], and [ar].
-  Uri? get envScript;
-
-  /// Arguments for [envScript].
-  List<String>? get envScriptArgs;
-
-  factory CCompilerConfig({
-    Uri? ar,
-    Uri? cc,
-    Uri? ld,
-    Uri? envScript,
-    List<String>? envScriptArgs,
-  }) = model.CCompilerConfig;
+  /// Constructs a config for a dry run by providing values for each field.
+  ///
+  /// `build.dart` hooks will most likely use [build] instead of this
+  /// constructor. However, for unit testing code which consumes a
+  /// [BuildConfig], this constructor facilitates easy construction.
+  ///
+  /// For the documentation of the parameters, see the equally named fields.
+  factory BuildConfig.dryRun({
+    required Uri outDir,
+    required String packageName,
+    required Uri packageRoot,
+    required OS targetOS,
+    required LinkModePreference linkModePreference,
+    Iterable<String>? supportedAssetTypes,
+  }) =>
+      BuildConfigImpl.dryRun(
+        outDir: outDir,
+        packageName: packageName,
+        packageRoot: packageRoot,
+        targetOS: targetOS as OSImpl,
+        linkModePreference: linkModePreference as LinkModePreferenceImpl,
+        supportedAssetTypes: supportedAssetTypes,
+      );
 }
