@@ -2,26 +2,25 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../api/build_mode.dart' as api;
+part of '../api/build_mode.dart';
 
-class BuildMode implements api.BuildMode {
+final class BuildModeImpl implements BuildMode {
   @override
   final String name;
 
-  const BuildMode._(this.name);
+  const BuildModeImpl._(this.name);
 
-  static const debug = BuildMode._('debug');
-  static const release = BuildMode._('release');
+  static const debug = BuildModeImpl._('debug');
+  static const release = BuildModeImpl._('release');
 
   static const values = [
     debug,
     release,
   ];
 
-  factory BuildMode.fromString(String target) =>
+  factory BuildModeImpl.fromString(String target) =>
       values.firstWhere((e) => e.name == target);
 
-  /// The `package:config` key preferably used.
   static const String configKey = 'build_mode';
 
   @override
