@@ -2,28 +2,24 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../api/ios_sdk.dart' as api;
+part of '../api/ios_sdk.dart';
 
-/// For an iOS target, a build is either done for the device or the simulator.
-///
-/// Only fat binaries or xcframeworks can contain both targets.
-class IOSSdk implements api.IOSSdk {
+final class IOSSdkImpl implements IOSSdk {
   final String xcodebuildSdk;
 
-  const IOSSdk._(this.xcodebuildSdk);
+  const IOSSdkImpl._(this.xcodebuildSdk);
 
-  static const iPhoneOs = IOSSdk._('iphoneos');
-  static const iPhoneSimulator = IOSSdk._('iphonesimulator');
+  static const iPhoneOS = IOSSdkImpl._('iphoneos');
+  static const iPhoneSimulator = IOSSdkImpl._('iphonesimulator');
 
   static const values = [
-    iPhoneOs,
+    iPhoneOS,
     iPhoneSimulator,
   ];
 
-  factory IOSSdk.fromString(String target) =>
+  factory IOSSdkImpl.fromString(String target) =>
       values.firstWhere((e) => e.xcodebuildSdk == target);
 
-  /// The `package:config` key preferably used.
   static const String configKey = 'target_ios_sdk';
 
   @override
