@@ -2,50 +2,31 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../api/link_mode_preference.dart' as api;
-import 'link_mode.dart';
+part of '../api/link_mode_preference.dart';
 
-class LinkModePreference implements api.LinkModePreference {
+final class LinkModePreferenceImpl implements LinkModePreference {
   @override
   final String name;
 
-  @override
-  final LinkMode preferredLinkMode;
+  const LinkModePreferenceImpl(this.name);
 
-  @override
-  final List<LinkMode> potentialLinkMode;
-
-  const LinkModePreference(
-    this.name, {
-    required this.preferredLinkMode,
-    required this.potentialLinkMode,
-  });
-
-  factory LinkModePreference.fromString(String name) =>
+  factory LinkModePreferenceImpl.fromString(String name) =>
       values.where((element) => element.name == name).first;
 
-  static const dynamic = LinkModePreference(
+  static const dynamic = LinkModePreferenceImpl(
     'dynamic',
-    preferredLinkMode: LinkMode.dynamic,
-    potentialLinkMode: [LinkMode.dynamic],
   );
 
-  static const static = LinkModePreference(
+  static const static = LinkModePreferenceImpl(
     'static',
-    preferredLinkMode: LinkMode.static,
-    potentialLinkMode: [LinkMode.static],
   );
 
-  static const preferDynamic = LinkModePreference(
+  static const preferDynamic = LinkModePreferenceImpl(
     'prefer-dynamic',
-    preferredLinkMode: LinkMode.dynamic,
-    potentialLinkMode: LinkMode.values,
   );
 
-  static const preferStatic = LinkModePreference(
+  static const preferStatic = LinkModePreferenceImpl(
     'prefer-static',
-    preferredLinkMode: LinkMode.static,
-    potentialLinkMode: LinkMode.values,
   );
 
   static const values = [
@@ -55,7 +36,6 @@ class LinkModePreference implements api.LinkModePreference {
     preferStatic,
   ];
 
-  /// The `package:config` key preferably used.
   static const String configKey = 'link_mode_preference';
 
   @override
