@@ -5,12 +5,10 @@
 import 'package:collection/collection.dart';
 import 'package:yaml/yaml.dart';
 
-import '../api/metadata.dart' as api;
 import '../utils/map.dart';
 import '../utils/yaml.dart';
 
-class Metadata implements api.Metadata {
-  @override
+class Metadata {
   final Map<String, Object> metadata;
 
   const Metadata(this.metadata);
