@@ -22,25 +22,28 @@ void main() {
     BuildOutput(
       timestamp: DateTime.parse('2022-11-10 13:25:01.000'),
       assets: [
-        Asset(
-          id: 'foo',
-          path: AssetAbsolutePath(Uri(path: 'path/to/libfoo.so')),
-          target: Target.androidX64,
-          linkMode: LinkMode.dynamic,
+        NativeCodeAsset(
+          package: 'my_package',
+          name: 'foo',
+          file: Uri(path: 'path/to/libfoo.so'),
+          linkMode: DynamicLoadingBundled(),
+          os: OS.android,
+          architecture: Architecture.x64,
         ),
-        Asset(
-          id: 'foo2',
-          path: AssetSystemPath(Uri(path: 'path/to/libfoo2.so')),
-          target: Target.androidX64,
-          linkMode: LinkMode.dynamic,
+        NativeCodeAsset(
+          package: 'my_package',
+          name: 'foo2',
+          linkMode: DynamicLoadingSystem(Uri(path: 'path/to/libfoo2.so')),
+          os: OS.android,
+          architecture: Architecture.x64,
         ),
       ],
-      dependencies: Dependencies([
+      dependencies: [
         Uri.file('path/to/file.ext'),
-      ]),
-      metadata: const Metadata({
+      ],
+      metadata: {
         'key': 'value',
-      }),
+      },
     );
   });
 }
